1.moved all the constants to a file 2.took util funcs to a util file 3.some time consts added to the contants 4.declared a helper function to createAndLabelNamespace if it doesnt Exist(the checking isnt in the func)
5.adjusted some logs so that they'd use the consts instead of the plain string. example:
from:
By("Create GPU Burn namespace 'test-gpu-burn'")
to:
By(fmt.Sprintf("Create GPU Burn namespace '%s'", gpuBurnNamespace))
